requirements.txt
llama_index==0.9.32
langchain==0.1.4
jieba
loguru
milvus==2.3.3
pymilvus==2.3.3
transformers
evaluate
text2vec
torch
elasticsearch
FlagEmbedding
rouge_score
datasets
pandas
numpy
tqdm
pandarallel
persist_to_disk
scikit-learn
sentence-transformers
matplotlib
selfcheckgpt
torchmetrics
nltk
pydantic
fuzzywuzzy